MacOS/rename_app.sh: resilience.

+++ otterlace/release/scripts/MacOS/rename_app.sh
@@ -4,8 +4,8 @@ set -e # bail out on error
 
 . "$( dirname "$0" )/_macos_above_app.sh" || exit 1
 
-old_app_name="$1"
-new_app_name="$2"
+old_app_name="${1%/}"           # strip trailing slash
+new_app_name="${2%/}"           # --"--
 
 if [ -z "${old_app_name}" ] || [ -z "${new_app_name}" ]; then
     echo "Usage: ${script_name} <old_app_name> <new_app_name>" >&2
@@ -14,6 +14,11 @@ fi
 
 old_non_dist_name="$( versioned_non_dist "${old_app_name}" )" || exit $?
 new_non_dist_name="${old_non_dist_name/${old_app_name}/${new_app_name}}" # retain versioning
+
+if [ -e "${new_app_name}" ]; then
+    echo "Destination '${new_app_name}' already exists."
+    exit 5
+fi
 
 mv -v "${old_app_name}" "${new_app_name}"
 mv -v "${non_dist}/${old_non_dist_name}" "${non_dist}/${new_non_dist_name}"
